% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_dosetable.R
\name{mod_server_dosetable}
\alias{mod_server_dosetable}
\title{Update Dosing Table}
\usage{
mod_server_dosetable(id)
}
\arguments{
\item{id}{Namespace}
}
\value{
Shiny server module
}
\description{
Update Dosing Table
}
