% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_mkdoseev.R
\name{mod_server_mkdoseev}
\alias{mod_server_mkdoseev}
\title{Dosing Server}
\usage{
mod_server_mkdoseev(id)
}
\arguments{
\item{id}{Namespace}
}
\value{
Shiny server module
}
\description{
Dosing Server
}
