% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_mkplot.R
\name{mod_server_mkplot}
\alias{mod_server_mkplot}
\title{Dosing Server}
\usage{
mod_server_mkplot(id, simdat)
}
\arguments{
\item{id}{Namespace}

\item{simdat}{simulation data.frame}
}
\value{
Shiny server module
}
\description{
Dosing Server
}
